server/ai/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.101.0
fasttext-wheel==0.9.2
gensim==4.3.1
h11==0.14.0
hazm==0.9.3
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
nltk==3.8.1
numpy==1.24.4
orjson==3.9.4
protobuf==4.21.12
pybind11==2.11.1
pydantic==2.1.1
pydantic-core==2.4.0
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
python-crfsuite==0.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
scikit-learn==1.3.0
scipy==1.10.1
smart-open==6.3.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing-extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
pandas==2.0.1
watchfiles==0.19.0
websockets==11.0.3
